+++ src/container.js
@@ -2,9 +2,9 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { cleanString } from '../lib/util'
 
-import HTML from './html'
-import Editor from './editor'
-import Text from './text'
+import HTML from './HTML'
+import Editor from './Editor'
+import Text from './Text'
 
 class Container extends React.Component {
   render () {
@@ -15,31 +15,31 @@ class Container extends React.Component {
     document.title = title
 
     return (
-      <div style={{ 'height': '100%', 'width': '100%' }}>
+      <div style={{ height: '100%', width: '100%' }}>
         <div className='navbar navbar-center'>
           <div className='container'>
-            <div className='navbar-title'><span className='text-black'>{ cleanString(title) }</span></div>
+            <div className='navbar-title'><span className='text-black'>{cleanString(title)}</span></div>
             <div className='nav'>
-              { Object.keys(nav).map((k, i) => {
-                return <a key={i} href={nav[k]} target='_blank' rel='noopener noreferrer'> { k } </a>
+              {Object.keys(nav).map((k, i) => {
+                return <a key={i} href={nav[k]} target='_blank' rel='noopener noreferrer'> {k} </a>
               })}
             </div>
           </div>
         </div>
         <div>
-          <h5 className='text-center description'> { cleanString(description) }</h5>
+          <h5 className='text-center description'> {cleanString(description)}</h5>
           <div style={{ width, margin: '0 auto' }}>
-            { body
+            {body
               ? body.map((block) => {
-                switch (block.type) {
-                  case 'code':
-                    return <Editor {...block} />
-                  case 'text':
-                    return <Text {...block} />
-                  case 'html':
-                    return <HTML {...block} />
-                }
-              })
+                  switch (block.type) {
+                    case 'code':
+                      return <Editor {...block} />
+                    case 'text':
+                      return <Text {...block} />
+                    case 'html':
+                      return <HTML {...block} />
+                  }
+                })
               : ''}
           </div>
         </div>
@@ -65,4 +65,4 @@ Container.defaultProps = {
   body: []
 }
 
-module.exports = Container
+export default Container