+++ src/editor.js
@@ -1,11 +1,9 @@
-import './editor.css'
-import 'krayon/dist/krayon.css'
+import './Editor.css'
 
-import Krayon from 'krayon'
 import React from 'react'
 import PropTypes from 'prop-types'
+import MonacoEditor from 'react-monaco-editor'
 
-import Textarea from './textarea.js'
 import { cleanString } from '../lib/util'
 
 class Editor extends React.Component {
@@ -18,12 +16,13 @@ class Editor extends React.Component {
       duration: 0
     }
   }
+
   run () {
     const self = this
     const { value } = this.state;
 
     (function () {
-      var c = {
+      const c = {
         cons: []
       }
       c.log = function () {
@@ -60,54 +59,59 @@ class Editor extends React.Component {
 
     return this
   }
+
   onChange (value) {
     this.setState({
       value
     })
   }
+
   render () {
     const { title, subtitle } = this.props
     const { value, output, duration } = this.state
     const { cons, val } = output
-    // Since we don't want to display all html entities we just want to show spans
-    // Since we don't want to add extra characters by making these entities (which would throw off the underlying layer) we use a full-width greater than and less than.
-    let colorized = Krayon(value.replace(/</g, '＜').replace(/>/g, '＞'))
 
-    return (<div className='editor'>
-      <div className='text-left text-black'>
-        { cleanString(title) }
-        { subtitle ? <div><small> { cleanString(subtitle) } </small></div> : '' }
-      </div>
-      <br />
-      <div style={{ position: 'relative' }}>
-        <Textarea
-          onChange={this.onChange.bind(this)}
-          value={value}
-        />
-        <pre className='textarea-overlay' style={{ position: 'absolute', top: 0, backgroundColor: 'rgba(255, 255, 255, 0)', whiteSpace: 'pre-wrap' }} dangerouslySetInnerHTML={{ __html: colorized }} />
-      </div>
-      <div style={{ overflow: 'auto', border: '1px solid #cfcfc4', padding: '0', borderBottomLeftRadius: '5px', borderBottomRightRadius: '5px' }}>
-        <div className='time'>
-         Run took { duration }ms
+    return (
+      <div className='editor'>
+        <div className='text-left text-black' style={{ marginBottom: '5px' }}>
+          {cleanString(title)}
+          {subtitle ? <div><small> {cleanString(subtitle)} </small></div> : ''}
         </div>
-        <div className='console'>
-          <span className='output'>
-            { (!val && !cons) ? <div><pre style={{ whiteSpace: 'pre-wrap', margin: 0, borderRadius: 0 }}>Output from the example appears here</pre></div> : '' }
-            { val ? <div><pre style={{ whiteSpace: 'pre-wrap', margin: 0, borderRadius: 0 }}>{val.toString()}</pre></div> : '' }
-            { cons && cons.length > 0
-              ? cons.map((c, i) => {
-                const { type, value } = c
-                if (type === 'html') {
-                  return <pre key={`${i}/${Date.now()}`} style={{ margin: '10px', border: '1px solid #f5f5f5', padding: '5px', position: 'relative' }}> <div dangerouslySetInnerHTML={{ __html: value.toString() }} /> </pre>
-                }
-                return <pre key={`${i}/${Date.now()}`} style={{ margin: '10px', border: '1px solid #f5f5f5', padding: '5px', position: 'relative' }}>{ value }</pre>
-              })
-              : '' }
-          </span>
-          <button className='run' type='button' onClick={this.run.bind(this)}>Run</button>
+
+        <div className='textarea' style={{ position: 'relative', border: '1px solid rgb(207, 207, 196)', borderBottom: 0, borderTopRightRadius: '5px', borderTopLeftRadius: '5px', padding: '6px' }}>
+          <MonacoEditor
+            width='100%'
+            height='200px'
+            language='javascript'
+            theme='vs'
+            value={value}
+            options={{ selectOnLineNumbers: true }}
+            onChange={this.onChange.bind(this)}
+          />
+        </div>
+        <div style={{ overflow: 'auto', border: '1px solid #cfcfc4', padding: '0', borderBottomLeftRadius: '5px', borderBottomRightRadius: '5px' }}>
+          <div className='time'>
+            Run took {duration}ms
+          </div>
+          <div className='console'>
+            <span className='output'>
+              {(!val && !cons) ? <div><pre style={{ whiteSpace: 'pre-wrap', margin: 0, borderRadius: 0 }}>Output from the example appears here</pre></div> : ''}
+              {val ? <div><pre style={{ whiteSpace: 'pre-wrap', margin: 0, borderRadius: 0 }}>{val.toString()}</pre></div> : ''}
+              {cons && cons.length > 0
+                ? cons.map((c, i) => {
+                    const { type, value } = c
+                    if (type === 'html') {
+                      return <pre key={`${i}/${Date.now()}`} style={{ margin: '10px', border: '1px solid #f5f5f5', padding: '5px', position: 'relative' }}> <div dangerouslySetInnerHTML={{ __html: value.toString() }} /> </pre>
+                    }
+                    return <pre key={`${i}/${Date.now()}`} style={{ margin: '10px', border: '1px solid #f5f5f5', padding: '5px', position: 'relative' }}>{value}</pre>
+                  })
+                : ''}
+            </span>
+            <button className='run' type='button' onClick={this.run.bind(this)}>Run</button>
+          </div>
         </div>
       </div>
-    </div>)
+    )
   }
 }
 
@@ -117,4 +121,4 @@ Editor.propTypes = {
   value: PropTypes.string
 }
 
-module.exports = Editor
+export default Editor