+++ src/index.js
@@ -1,5 +1,0 @@
-module.exports = {
-  HTML: require('./html'),
-  Text: require('./text'),
-  Editor: require('./editor')
-}