+++ src/templates/landing.js
@@ -4,7 +4,8 @@ import './landing.css'
 import React from 'react'
 import PropTypes from 'prop-types'
 import { render } from 'react-dom'
-import { HTML } from '../'
+
+import HTML from '../HTML'
 import { cleanString } from '../../lib/util'
 
 class Landing extends React.Component {
@@ -19,10 +20,10 @@ class Landing extends React.Component {
       <div id='container' style={{ width, textAlign: 'center' }}>
         <div className='navbar'>
           <div className='container'>
-            <div className='navbar-title'><span className='text-black'>{ cleanString(title) }</span></div>
+            <div className='navbar-title'><span className='text-black'>{cleanString(title)}</span></div>
             <div className='nav'>
-              { Object.keys(nav).map((k, i) => {
-                return <a key={i} href={nav[k]} target='_blank' rel='noopener noreferrer'> { k } </a>
+              {Object.keys(nav).map((k, i) => {
+                return <a key={i} href={nav[k]} target='_blank' rel='noopener noreferrer'> {k} </a>
               })}
             </div>
           </div>
@@ -69,7 +70,7 @@ if ((window && window.config) || global.config) {
     const hash = injectedConfig.hash
 
     setInterval(function () {
-      var xhttp = new XMLHttpRequest()
+      const xhttp = new XMLHttpRequest()
       xhttp.onreadystatechange = function () {
         if (this.readyState === 4 && this.status === 200) {
           const response = JSON.parse(xhttp.responseText)