+++ src/templates/readme.js
@@ -8,8 +8,8 @@ import { render } from 'react-dom'
 import Markdown from 'react-markdown'
 import { cleanString } from '../../lib/util'
 
-import HTML from '../html'
-import Editor from '../editor'
+import HTML from '../HTML'
+import Editor from '../Editor'
 
 class Readme extends React.Component {
   render () {
@@ -23,26 +23,29 @@ class Readme extends React.Component {
       <div id='container'>
         <div className='navbar navbar-center'>
           <div className='container'>
-            <div className='navbar-title'><span className='text-black'>{ cleanString(title) }</span></div>
+            <div className='navbar-title'><span className='text-black'>{cleanString(title)}</span></div>
             <div className='nav'>
-              { Object.keys(nav).map((k, i) => {
-                return <a key={i} href={nav[k]} target='_blank' rel='noopener noreferrer'> { k } </a>
+              {Object.keys(nav).map((k, i) => {
+                return <a key={i} href={nav[k]} target='_blank' rel='noopener noreferrer'> {k} </a>
               }, [])}
             </div>
           </div>
         </div>
         <div id='container-content'>
           <div style={{ width, margin: '0 auto' }}>
-            <Markdown source={readme} renderers={{
-              'code': function (prop) {
-                if (prop.language === 'javascript') {
-                  return <Editor value={prop.value} />
+            <Markdown
+              source={readme} renderers={{
+                code: function (prop) {
+                  if (prop.language === 'javascript') {
+                    return <Editor value={prop.value} />
+                  }
+                  return (
+                    <pre className='code'>
+                      <code> {prop.value} </code>
+                    </pre>
+                  )
                 }
-                return <pre className='code'>
-                  <code> { prop.value } </code>
-                </pre>
-              }
-            }}
+              }}
             />
           </div>
         </div>
@@ -83,7 +86,7 @@ if ((window && window.config) || global.config) {
     const hash = injectedConfig.hash
 
     setInterval(function () {
-      var xhttp = new XMLHttpRequest()
+      const xhttp = new XMLHttpRequest()
       xhttp.onreadystatechange = function () {
         if (this.readyState === 4 && this.status === 200) {
           const response = JSON.parse(xhttp.responseText)